Player_t: use int for VineNPC and VineBGO

+++ src/globals.h
@@ -541,9 +541,9 @@ struct Player_t
 //    GrabSpeed As Single
     float GrabSpeed = 0.0f;
 //    VineNPC As Double 'the NPC that the player is climbing
-    double VineNPC = 0.0;
+    int VineNPC = 0;
 //  EXTRA:  Fence BGO
-    double VineBGO = 0.0;
+    int VineBGO = 0;
 //    Wet As Integer 'weather or not the player is under water
     int Wet = 0;
 //    WetFrame As Boolean 'true if the play should be swimming
